+++ server/model/common/request/common.go
@@ -9,7 +9,7 @@ type PageInfo struct {
 
 // GetById Find by id structure
 type GetById struct {
-	ID int `json:"id" form:"id"` // 主键ID
+	ID uint64 `json:"id" form:"id"` // 主键ID
 }
 
 func (r *GetById) Uint() uint {
