+++ server/model/system/request/sys_authority_btn.go
@@ -1,7 +1,7 @@
 package request
 
 type SysAuthorityBtnReq struct {
-	MenuID      uint   `json:"menuID"`
+	MenuID      uint64 `json:"menuID"`
 	AuthorityId uint   `json:"authorityId"`
 	Selected    []uint `json:"selected"`
 }
