+++ server/service/system/sys_auto_code.go
@@ -254,7 +254,7 @@ func makeDictTypes(autoCode *system.AutoCodeStruct) {
 // @param: model.AutoCodeStruct
 // @return: err error
 
-func (autoCodeService *AutoCodeService) CreateTemp(autoCode system.AutoCodeStruct, ids ...uint) (err error) {
+func (autoCodeService *AutoCodeService) CreateTemp(autoCode system.AutoCodeStruct, ids ...uint64) (err error) {
 	makeDictTypes(&autoCode)
 	for i := range autoCode.Fields {
 		if autoCode.Fields[i].FieldType == "time.Time" {
@@ -515,7 +515,7 @@ func (autoCodeService *AutoCodeService) addAutoMoveFile(data *tplData) {
 // @param: a *model.AutoCodeStruct
 // @return: err error
 
-func (autoCodeService *AutoCodeService) AutoCreateApi(a *system.AutoCodeStruct) (ids []uint, err error) {
+func (autoCodeService *AutoCodeService) AutoCreateApi(a *system.AutoCodeStruct) (ids []uint64, err error) {
 	apiList := []system.SysApi{
 		{
 			Path:        "/" + a.Abbreviation + "/" + "create" + a.StructName,
